requirements.txt
streamlit>=1.24.0
openai>=1.0.0
python-dotenv
